剑指Offer--038-数字在排序数组中出现的次数--http://blog.csdn.net/gatieme/article/details/51335323

+++ 038-数字在排序数组中出现的次数/02-binarysearch.cpp
@@ -1,7 +1,9 @@
 #include <iostream>
 #include <vector>
 
+
 using namespace std;
+
 
 //  调试开关
 #define __tmain main
@@ -16,9 +18,12 @@ using namespace std;
 
 #endif // __tmain
 
+
+
 /* 先用二分查找找出某个k出现的位置，然后再分别向前和向后查找总的个数*/
 class Solution
 {
+
 public:
     int GetNumberOfK(vector<int> data,int key)
     {
@@ -30,11 +35,11 @@ public:
         //  用二分查找查找到Key的位置
         int index = BinarySearch(data, 0, data.size( ) - 1, key);
         //int index = BinarySearch(data, key);
-        if(index == -1) 
+        if(index == -1)
         {
             return 0;
         }
-        
+
         int count = 1;
         // 查找前面部分Key的个数
         for(int j = index - 1;
@@ -47,7 +52,7 @@ public:
 
         //  查找后面部分Key的个数
         for(int j = index + 1;
-            j < data.size( ) && data[j] == key;
+            j < (int)data.size( ) && data[j] == key;
             j++)
         {
             count++;
@@ -55,14 +60,14 @@ public:
 
         return count;
     }
-    
+
     int BinarySearch(vector<int> &data, int begin, int end ,int key)
     {
         if(begin > end)
         {
             return -1;
         }
-        
+
         int mid = (begin + end) / 2;
         debug <<"mid = " <<mid <<", data = " <<data[mid] <<endl;
         if(data[mid] == key)
@@ -85,29 +90,29 @@ public:
     {
         int low = 0, high = array.size( ) - 1;
 
-        int mid;  
-        
-        while (low <= high)  
-        {  
-            mid = (low + high) / 2;  
-            
+        int mid;
+
+        while (low <= high)
+        {
+            mid = (low + high) / 2;
+
             if(key == array[mid])
-            {  
-                return mid;  
+            {
+                return mid;
             }
-            
+
             if(key < array[mid])
             {
-                high = mid - 1;  
+                high = mid - 1;
             }
 
             if(key > array[mid])
             {
-                low = mid + 1;  
+                low = mid + 1;
             }
         }
 
-        return -1;  
+        return -1;
     }
 };
